Move `Utils` to the `IndicoToolkit` namespace with generic exceptions
